move the invoice pipes to appropriate directory

Took 53 seconds
